refactor: Card 컴포넌트 Common으로 분리
