+++ frontend/src/components/Common/Card/Card.tsx
@@ -2,7 +2,7 @@ import { css, cx } from '@style/css';
 
 import { HTMLAttributes } from 'react';
 
-interface Props extends HTMLAttributes<HTMLDivElement> {}
+export interface Props extends HTMLAttributes<HTMLDivElement> {}
 
 export function Card({ className, children, ...props }: Props) {
   return (
