+++ frontend/src/components/Common/Card/index.ts
@@ -1,0 +1,2 @@
+export { Card } from './Card';
+export type { Props as CardProps } from './Card';
