+++ frontend/src/components/CompetitionDetail/DuringCompetition.tsx
@@ -3,12 +3,12 @@ import { css, cx } from '@style/css';
 import { HTMLAttributes } from 'react';
 
 import { CompetitionInfo } from '@/apis/competitions';
+import { Card } from '@/components/Common';
 import { formatDate } from '@/utils/date';
 
 import { Chip, Space, Text, VStack } from '../Common';
 import ViewDashboardButton from '../Main/Buttons/ViewDashboardButton';
 import EnterCompetitionButton from './Buttons/EnterCompetitionButton';
-import { Card } from './Card';
 import CompetitionDetailInfo from './CompetitionDetailInfo';
 import CompetitionMembersInfo from './CompetitionMembersInfo';
 import { buttonContainerStyle } from './styles/styles';
